% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_quarantine_app}
\alias{run_quarantine_app}
\title{This function launches the interactive Shiny app included with the
`covidQuarantineViz` package.}
\usage{
run_quarantine_app()
}
\description{
This function launches the interactive Shiny app included with the
`covidQuarantineViz` package.
}
\examples{
# run_quarantine_app()
}
